+++ market/rsx_ssot.go
@@ -41,4 +41,5 @@ func (a *Frame) rsxScanConfigLocked() indicators.RSXScanConfig {
 
 func (a *Frame) rebuildRSXAnnotationsLocked() {
 	a.Annotations = a.Annotations[:0]
+	a.rebuildRSTVFactsLocked()
 }